Fix request loop unblocking during continuous inflood of votes
In case there is an continuous inflood of votes, is_active often does not equal false long enough to unblock the request loop thread which is then stuck in flush_active for multiple seconds or even minutes. This can be solved by only checking once for is_active in flush_active and unblocking the request loop thread by a separate flush notification.